+++ wgpu-hal/src/metal/adapter.rs
@@ -308,7 +308,7 @@ impl crate::Adapter<super::Api> for super::Adapter {
             },
             composite_alpha_modes: vec![
                 wgt::CompositeAlphaMode::Opaque,
-                wgt::CompositeAlphaMode::PreMultiplied,
+                wgt::CompositeAlphaMode::PostMultiplied,
             ],
 
             current_extent,