+++ wgpu/src/backend/web.rs
@@ -1213,12 +1213,20 @@ impl crate::Context for Context {
         ]
     }
 
-    fn surface_get_supported_modes(
+    fn surface_get_supported_present_modes(
         &self,
         _surface: &Self::SurfaceId,
         _adapter: &Self::AdapterId,
     ) -> Vec<wgt::PresentMode> {
         vec![wgt::PresentMode::Fifo]
+    }
+
+    fn surface_get_supported_alpha_modes(
+        &self,
+        _surface: &Self::SurfaceId,
+        _adapter: &Self::AdapterId,
+    ) -> Vec<wgt::CompositeAlphaMode> {
+        vec![wgt::CompositeAlphaMode::Opaque]
     }
 
     fn surface_configure(
@@ -1229,6 +1237,9 @@ impl crate::Context for Context {
     ) {
         if let wgt::PresentMode::Mailbox | wgt::PresentMode::Immediate = config.present_mode {
             panic!("Only FIFO/Auto* is supported on web");
+        }
+        if config.alpha_mode != wgt::CompositeAlphaMode::Opaque {
+            panic!("Only Opaque alpha mode is supported on web");
         }
         let mut mapped =
             web_sys::GpuCanvasConfiguration::new(&device.0, map_texture_format(config.format));